@@ -115,8 +115,8 @@
   "dependencies": {
     "@craco/craco": "5.6.4",
     "@electron/remote": "2.0.7",
-    "@material-ui/core": "4.10.2",
-    "@material-ui/icons": "3.0.2",
+    "@material-ui/core": "4.12.3",
+    "@material-ui/icons": "4.11.2",
     "@neutrinog/electron-dl": "1.11.1",
     "@sendgrid/mail": "6.4.0",
     "adm-zip": "0.4.13",
@@ -146,7 +146,7 @@
     "ospath": "1.2.2",
     "path-extra": "3.0.0",
     "prop-types": "15.7.2",
-    "react": "16.8.6",
+    "react": "16.14.0",
     "react-bootstrap": "0.32.4",
     "react-container-dimensions": "1.3.3",
     "react-dnd": "2.5.4",
@@ -155,7 +155,7 @@
     "react-localize-redux": "3.5.3",
     "react-redux": "5.1.1",
     "react-remarkable": "1.1.3",
-    "react-scripts": "3.4.1",
+    "react-scripts": "3.4.4",
     "react-select": "1.2.0",
     "react-spotlight-clickable": "1.0.5",
     "react-toggle": "4.0.2",
@@ -176,7 +176,7 @@
     "string-punctuation-tokenizer": "^2.2.0",
     "sudo-prompt": "6.2.1",
     "tc-electron-env": "0.10.0",
-    "tc-source-content-updater": "1.2.6",
+    "tc-source-content-updater": "1.2.8",
     "tc-strings": "0.1.7",
     "tc-tool": "4.1.0",
     "tc-ui-toolkit": "6.0.4",
