@@ -70,7 +70,7 @@ const ToolsCards = ({
           tools.map((tool, i) => {
             const glSelected = resourcesHelpers.splitVersionAndOwner(manifest.toolsSelectedGLs?.[tool.name] || '').version;
             const glOwnerSelected = manifest.toolsSelectedOwners?.[tool.name] || DEFAULT_OWNER;
-            const availableCategories = getAvailableCategories(glSelected, tool.name, projectSaveLocation, { withCategoryName: true });
+            const availableCategories = getAvailableCategories(glSelected, tool.name, projectSaveLocation, { withCategoryName: true }, glOwnerSelected);
             let isOLBookVersionMissing = false;
             let missingOLResource = {};
 
