@@ -27,27 +27,33 @@ function setDotEnv(dotenvConfig, app) {
 export function initEnv(msg) {
   console.log(`initEnv - initial is electron: ${env.isElectron()}`);
   const { app } = require('@electron/remote');
-  const dotenv = require('dotenv');
   let failed = true;
+  let dotenvConfig;
+  const pe = process.env;
+  const build = pe.BUILD;
 
-  // const fs = require('fs-extra');
-  // console.log('folder', __dirname);
-  //
-  // const folders = [__dirname, path.join(__dirname, 'static'), path.join(__dirname, '..')];
-  //
-  // for (const folder of folders) {
-  //   try {
-  //     const files = fs.readdirSync(folder);
-  //     console.log(`files ${folder}`, files);
-  //     // eslint-disable-next-line no-empty
-  //   } catch (e) {
-  //     console.warn(`could not read files from ${folder}`, e);
-  //   }
-  // }
+  console.log('folder', __dirname);
 
-  const configPath = path.join(__dirname, 'cfg.txt');
-  const envConfig = dotenv.config({ path: configPath });
-  const dotenvConfig = envConfig?.parsed;
+  const folders = [__dirname, path.join(__dirname, 'static'), path.join(__dirname, '..')];
+  const fs = require('fs-extra');
+
+  for (const folder of folders) {
+    try {
+      const files = fs.readdirSync(folder);
+      console.log(`files ${folder}`, files);
+      // eslint-disable-next-line no-empty
+    } catch (e) {
+      console.warn(`could not read files from ${folder}`, e);
+    }
+  }
+
+  if (build) {
+    dotenvConfig = {}; // if environment is set, don't need additional env variables
+  } else { // need to load env variables
+    const dotenv = require('dotenv');
+    const configPath = path.join(__dirname, 'cfg.txt');
+    dotenvConfig = dotenv.config({ path: configPath })?.parsed;
+  }
 
   if (!dotenvConfig) {
     if (app) {
